move requests.get mock to conftest.py
