+++ tests/conftest.py
@@ -1,0 +1,16 @@
+"""Package-wide test fixtures."""
+from unittest.mock import Mock
+
+import pytest
+from pytest_mock import MockFixture
+
+
+@pytest.fixture
+def mock_requests_get(mocker: MockFixture) -> Mock:
+    """Fixture for mocking requests.get."""
+    mock = mocker.patch("requests.get")
+    mock.return_value.__enter__.return_value.json.return_value = {
+        "title": "Lorem Ipsum",
+        "extract": "Lorem ipsum dolor sit amet",
+    }
+    return mock
